Update with PR comments
