+++ library-no-op/api/library-no-op.api
@@ -8,7 +8,6 @@ public final class com/chuckerteam/chucker/api/Chucker {
 	public static final fun dismissNotifications (Landroid/content/Context;)V
 	public static final fun getLaunchIntent (Landroid/content/Context;)Landroid/content/Intent;
 	public final fun isOp ()Z
-	public static final fun writeTransactions (Landroid/content/Context;Ljava/lang/Long;Ljava/lang/Long;)Landroid/net/Uri;
 }
 
 public final class com/chuckerteam/chucker/api/ChuckerCollector {
@@ -18,6 +17,7 @@ public final class com/chuckerteam/chucker/api/ChuckerCollector {
 	public synthetic fun <init> (Landroid/content/Context;ZLcom/chuckerteam/chucker/api/RetentionManager$Period;ILkotlin/jvm/internal/DefaultConstructorMarker;)V
 	public final fun getShowNotification ()Z
 	public final fun setShowNotification (Z)V
+	public final fun writeTransactions (Landroid/content/Context;Ljava/lang/Long;Ljava/lang/Long;Lkotlin/coroutines/Continuation;)Ljava/lang/Object;
 }
 
 public final class com/chuckerteam/chucker/api/ChuckerInterceptor : okhttp3/Interceptor {
