+++ library-no-op/src/main/kotlin/com/chuckerteam/chucker/api/ChuckerCollector.kt
@@ -1,6 +1,7 @@
 package com.chuckerteam.chucker.api
 
 import android.content.Context
+import android.net.Uri
 
 /**
  * No-op implementation.
@@ -10,4 +11,10 @@ public class ChuckerCollector @JvmOverloads constructor(
     context: Context,
     public var showNotification: Boolean = true,
     retentionPeriod: RetentionManager.Period = RetentionManager.Period.ONE_WEEK
-)
+) {
+    public suspend fun writeTransactions(
+        context: Context,
+        maxTransactions: Long?,
+        startTimestamp: Long?,
+    ): Uri? = null
+}
